update tests to use neon and clean up layer provides
